+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
 
         </activity>
         <service
-            android:name=".data.Notifications"
+            android:name=".notifications.NotificationClient"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
