+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/notifications/NotificationClient.kt
@@ -1,4 +1,4 @@
-package io.github.viabachelora23michaelkutaibakasper.bprapp.data
+package io.github.viabachelora23michaelkutaibakasper.bprapp.notifications
 
 import android.util.Log
 import androidx.compose.runtime.mutableStateOf
@@ -7,7 +7,7 @@ import com.google.firebase.ktx.Firebase
 import com.google.firebase.messaging.FirebaseMessagingService
 import io.github.viabachelora23michaelkutaibakasper.bprapp.data.sign_in.FireStoreClient
 
-class Notifications() : FirebaseMessagingService() {
+class NotificationClient() : FirebaseMessagingService() {
     var user = mutableStateOf(Firebase.auth.currentUser)
 
     override fun onNewToken(token: String) {
@@ -15,7 +15,7 @@ class Notifications() : FirebaseMessagingService() {
         // If you want to send messages to this application instance or
         // manage this apps subscriptions on the server side, send the
         // FCM registration token to your app server.
-        FireStoreClient().UpdateFirebaseMessagingToken(token, user.value!!.uid)
+        FireStoreClient().updateFirebaseMessagingToken(token, user.value!!.uid)
     }
 
 }
